+++ src/Router.js
@@ -1,7 +1,8 @@
 import React from 'react';
-import { Scene, Router, } from 'react-native-router-flux';
+import { Scene, Router, Actions } from 'react-native-router-flux';
 import LoginForm from './components/LoginForm';
 import EmployeeList from './components/EmployeeList';
+import EmployeeCreate from './components/EmployeeCreate';
 
 const RouterComponent = () => {
     return (
@@ -13,12 +14,14 @@ const RouterComponent = () => {
                 <Scene key="main">
                     <Scene
                         rightTitle="Add"
-                        onRight={() => {console.log('alert')}}
+                        onRight={() => Actions.employeeCreate()}
                         key="employeeList"
                         sceneStyle={styles.formStyle}
                         component={EmployeeList}
                         title="Employee List"
+                        initial
                     />
+                    <Scene key="employeeCreate" sceneStyle={styles.formStyle} component={EmployeeCreate} title="Create Employee" />
                 </Scene>
             </Scene>
         </Router>
